+++ cypress/pages/index.ts
@@ -1,6 +1,5 @@
-export {LoginPage} from '../pages/homePage';
-export {CategoriesPage} from '../pages/shoppingPage';
-export {CartPage} from '../pages/cartPage';
-export {PurchasePage} from '../pages/purchasePage';
-export {Contact} from '../pages/contactPage';
+export {LoginPage} from './loginPage';
+export {ShoppingPage} from '../pages/shoppingPage';
+
+
