+++ cypress/pages/shoppingPage.ts
@@ -1,33 +1,30 @@
 const products = Cypress.env("Products");
 
-class CategoriesPage {
+class ShoppingPage {
+
+  private elementSelector
+  private navBarSelector
+  private categoriesSelector
+  private cartValues
+  private cart
+
   constructor(){
     this.elementSelector = ".card-title",
-    this.containerElement = "#tbodyid",
-    this.addToCart = ".btn-lg",
-    this.addToCartBtn = "Add to cart"
     this.navBarSelector = ".nav-link",
-    this.home = "Home",
     this.categoriesSelector = ".list-group-item",
     this.cartValues = "table tbody tr td",
     this.cart = "Cart"
   }
 
-  addElementToCart(elements){
-    elements.forEach(element => {
-      cy.get(this.categoriesSelector).contains(element.category).click();
-      cy.get(this.elementSelector).contains(element.name).click();
-      cy.get(this.addToCart).contains(this.addToCartBtn).click();
-      cy.get(this.navBarSelector).contains(this.home).click();
-    });
+  addElementToCart(element) {
+    cy.get(this.categoriesSelector).contains(element.category).click();
+    cy.get(this.elementSelector).contains(element.name).click();
   }
-
-  verifyTitleElement(elements) {
+  
+  verifyElementAdded(element) {
     cy.get(this.navBarSelector).contains(this.cart).click();
-    elements.forEach(element => {
-      cy.get(this.cartValues).should("contain", element.name);
-    });
+    cy.get(this.cartValues).should("contain", element.name);
   }
 }
 
-export {CategoriesPage};
+export {ShoppingPage};
