+++ cypress/support/commands.ts
@@ -1,4 +1,5 @@
 // ***********************************************
+import 'cypress-xpath';
 // This example commands.js shows you how to
 // create various custom commands and overwrite
 // existing commands.
