+++ package.json
@@ -27,6 +27,7 @@
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^6.7.3",
     "cypress": "^13.2.0",
+    "cypress-xpath": "^2.0.1",
     "eslint": "^8.50.0",
     "eslint-config-standard-with-typescript": "^39.1.0",
     "eslint-plugin-cypress": "^2.15.1",
